refactor: replace AuthModal with AuthDialog and update MainMenu
- Replaced AuthModal with a new AuthDialog component for improved functionality and clarity.
- Updated ExcalidrawWrapper to use AuthDialog, simplifying the authentication flow.
- Removed the old AuthModal file.
